@@ -13,10 +13,10 @@ publish_to: 'none'
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.1+32
+version: 1.0.3+43
 
 environment:
-  sdk: ">=2.14.4 <3.0.0"
+  sdk: ^2.17.1
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -32,20 +32,18 @@ dependencies:
 
   cupertino_icons: ^1.0.4
   audioplayers: ^0.20.1
-  sqflite: ^2.0.2+1
-  animations: ^2.0.2
+  animations: ^2.0.3
   event_bus: ^2.0.0
-  flutter_platform_widgets: ^1.20.0
   package_info_plus: ^1.4.2
-  flutter_svg: ^1.0.3
+  flutter_svg: ^1.1.0
   shared_preferences: ^2.0.15
-  google_sign_in: ^5.3.1
+  google_sign_in: ^5.3.2
   sign_in_with_apple: ^4.0.0
   crypto: ^3.0.2
   flutter_dotenv: ^5.0.2
-  flutter_slidable: ^1.2.1
+  flutter_slidable: ^1.3.0
   url_launcher: ^6.1.2
-  camera: ^0.9.6
+  camera: ^0.9.7+1
   flutter_linkify: ^5.0.2
   path_provider: ^2.0.10
   flutter_exif_rotation: ^0.5.1
@@ -60,20 +58,23 @@ dependencies:
   share_plus: ^4.0.4
   http: ^0.13.4
   intl: ^0.17.0
+  isar: ^2.5.0
+  isar_flutter_libs: ^2.5.0
+
 
   googleapis: ^8.1.0
   googleapis_auth: ^1.3.1
   extension_google_sign_in_as_googleapis_auth: ^2.0.4
   google_mobile_ads: ^1.2.0
 
-  firebase_core: ^1.17.0
-  firebase_crashlytics: ^2.8.0
-  firebase_auth: ^3.3.18
-  cloud_firestore: ^3.1.15
-  firebase_analytics: ^9.1.8
-  firebase_dynamic_links: ^4.2.3
-  firebase_performance: ^0.8.0+12
-  cloud_functions: ^3.2.15
+  firebase_core: ^1.17.1
+  firebase_crashlytics: ^2.8.1
+  firebase_auth: ^3.3.19
+  cloud_firestore: ^3.1.17
+  firebase_analytics: ^9.1.9
+  firebase_dynamic_links: ^4.2.5
+  firebase_performance: ^0.8.0+13
+  cloud_functions: ^3.2.16
 
 dev_dependencies:
   flutter_test:
@@ -85,6 +86,8 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^2.0.1
+  isar_generator: ^2.5.0
+  build_runner: any
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
