@@ -2,16 +2,17 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:cloud_functions/cloud_functions.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:submon/db/digestive.dart';
-import 'package:submon/db/memorize_card_folder.dart';
+import 'package:submon/batch_operation.dart';
 import 'package:submon/db/shared_prefs.dart';
-import 'package:submon/db/sql_provider.dart';
-import 'package:submon/db/submission.dart';
-import 'package:submon/db/timetable.dart';
-import 'package:submon/db/timetable_class_time.dart';
-import 'package:submon/db/timetable_table.dart';
-import 'package:submon/events.dart';
+import 'package:submon/isar_db/isar_digestive.dart';
+import 'package:submon/isar_db/isar_provider.dart';
+import 'package:submon/isar_db/isar_submission.dart';
+import 'package:submon/isar_db/isar_timetable.dart';
+import 'package:submon/isar_db/isar_timetable_class_time.dart';
+import 'package:submon/isar_db/isar_timetable_table.dart';
+import 'package:submon/main.dart';
 import 'package:submon/utils/firestore.dart';
+import 'package:submon/utils/ui.dart';
 
 class FirestoreProvider {
   FirestoreProvider(this.collectionId);
@@ -38,6 +39,11 @@ class FirestoreProvider {
   Future<QuerySnapshot<Map<String, dynamic>>> get() async {
     assert(userDoc != null);
     return await userDoc!.collection(collectionId).get();
+  }
+
+  Future<DocumentSnapshot<Map<String, dynamic>>> getDoc(String id) async {
+    assert(userDoc != null);
+    return await userDoc!.collection(collectionId).doc(id).get();
   }
 
   Future<bool?> exists(String docId) async {
@@ -90,16 +96,18 @@ class FirestoreProvider {
     }
   }
 
-  static Future<void> setTimetableNotification(
-      TimetableNotification data) async {
+  static Future<void> setTimetableNotificationTime(TimeOfDay? time) async {
     if (userDoc != null) {
-      String? timeString;
-      if (data.time != null) {
-        timeString = "${data.time!.hour}:${data.time!.minute}";
-      }
       await userDoc!.set({
-        "timetableNotificationTime": timeString,
-        "timetableNotificationId": data.id
+        "timetableNotificationTime": time != null ? "${time.hour}:${time.minute}" : null,
+      }, SetOptions(merge: true));
+    }
+  }
+
+  static Future<void> setTimetableNotificationId(int id) async {
+    if (userDoc != null) {
+      await userDoc!.set({
+        "timetableNotificationId": id,
       }, SetOptions(merge: true));
     }
   }
@@ -154,10 +162,11 @@ class FirestoreProvider {
     }, SetOptions(merge: true));
   }
 
-  static Future<ConfigData?> get config async {
+  static Future<UserConfig?> get config async {
     if (userDoc != null) {
-      var snapshot = await userDoc!.get();
-      return ConfigData.fromMap(snapshot.data() as Map<String, dynamic>);
+      return (await userDoc!
+          .withConverter<UserConfig>(fromFirestore: UserConfig.fromFirestore,
+          toFirestore: (config, options) => config.toFirestore()).get()).data();
     } else {
       return null;
     }
@@ -167,91 +176,196 @@ class FirestoreProvider {
     await FirebaseFunctions.instanceFor(region: "asia-northeast1")
         .httpsCallable("createUser")
         .call();
-    await userDoc!.set({"schemaVersion": schemaVer});
+    await userDoc!.set({"schemaVersion": schemaVersion});
   }
 
-  static Future<void> checkMigration() async {
-    if (userDoc == null) return;
+  static Future<bool> checkMigration() async {
+    if (userDoc == null) return false;
     var snapshot = await userDoc!.get();
 
-    var schemaVersion = (snapshot.data() as dynamic)?["schemaVersion"];
+    var serverSchemaVersion = (snapshot.data() as dynamic)?["schemaVersion"];
 
-    if (schemaVersion == null) {
-      await userDoc!.set({"schemaVersion": schemaVer}, SetOptions(merge: true));
+    if (serverSchemaVersion == null) {
+      await userDoc!
+          .set({"schemaVersion": schemaVersion}, SetOptions(merge: true));
+      return false;
     } else {
-      if (schemaVersion < schemaVer) {
+      if (serverSchemaVersion < schemaVersion) {
+        showLoadingModal(globalContext!);
+
         // migrate (server side?)
+        await _migrate(serverSchemaVersion, snapshot.data() as dynamic);
+
         await userDoc!
-            .set({"schemaVersion": schemaVer}, SetOptions(merge: true));
-      } else if (schemaVersion > schemaVer) {
-        throw SchemaVersionMismatchException(schemaVer, schemaVersion);
+            .set({"schemaVersion": schemaVersion}, SetOptions(merge: true));
+
+        Navigator.of(globalContext!, rootNavigator: true).pop();
+
+        return true;
+      } else if (serverSchemaVersion > schemaVersion) {
+        throw SchemaVersionMismatchException(
+            serverSchemaVersion, schemaVersion);
       }
+      return false;
     }
+  }
+
+  static Future<void> _migrate(int oldVersion, Map<String, dynamic> userConfig) async {
+    if (oldVersion == 4) {
+      var operations = <BatchOperation>[];
+
+      if (userConfig["timetableNotificationId"] == null) {
+        operations.add(BatchOperation.set(doc: userDoc!, data: {
+          "timetableNotificationId": -1
+        }, setOptions: SetOptions(merge: true)));
+      }
+
+      var submissions = await submission.get();
+      for (var item in submissions.docs) {
+        var data = item.data();
+        data["details"] = data["detail"];
+        data["due"] = data["date"];
+        data["done"] = data["done"] == 1;
+        data["important"] = data["important"] == 1;
+        data.remove("detail");
+        data.remove("date");
+        operations.add(BatchOperation.set(
+          doc: userDoc!.collection("submission").doc(item.id),
+          data: data,
+        ));
+      }
+
+      var digestives = await digestive.get();
+      for (var item in digestives.docs) {
+        operations.add(BatchOperation.set(
+          doc: item.reference,
+          data: {
+            "done": item.data()["done"] == 1,
+          },
+          setOptions: SetOptions(merge: true),
+        ));
+      }
+
+      var mainTimetable = await timetable.getDoc("main");
+      if (mainTimetable.exists) {
+        var data = mainTimetable.data()!;
+        if (data["cells"] != null) {
+          data["cells"] = (data["cells"] as Map<String, dynamic>)
+              .map((key, value) => MapEntry(key, value..["tableId"] = -1));
+        }
+        operations.add(BatchOperation.set(
+          doc: userDoc!.collection("timetable").doc("-1"),
+          data: data,
+        ));
+        operations.add(BatchOperation.delete(
+          doc: userDoc!.collection("timetable").doc("main"),
+        ));
+      }
+
+      var timetableClassTimes = await timetableClassTime.get();
+      for (var item in timetableClassTimes.docs) {
+        var data = item.data();
+        data["period"] = data["id"];
+        data.remove("id");
+
+        operations.add(BatchOperation.set(
+            doc: userDoc!.collection("timetableClassTime").doc(item.id),
+            data: data));
+      }
+
+      await BatchOperation.commit(operations);
+
+      oldVersion++;
+    }
+  }
+
+  static void setLastAppOpenedToCurrentTime() {
+    userDoc?.set({"lastAppOpened": FieldValue.serverTimestamp()},
+        SetOptions(merge: true));
   }
 
   ///
   /// if value is changed, true will be returned.
   ///
-  static Future<bool> fetchData(
-      {required BuildContext context, bool force = false}) async {
-    await FirestoreProvider.checkMigration();
-
+  static Future<bool> fetchData({bool force = false}) async {
     var changed = !force ? await FirestoreProvider.checkTimestamp() : true;
+
+    if (await FirestoreProvider.checkMigration()) {
+      changed = true;
+    }
+
+    setLastAppOpenedToCurrentTime();
 
     if (changed == true) {
       final configData = await config;
       final submissionSnapshot = await submission.get();
       final digestiveSnapshot = await digestive.get();
-      final timetableDataSnapshot = await timetable.get();
+      final timetableSnapshot = await timetable.get();
       final timetableClassTimeDataSnapshot = await timetableClassTime.get();
       final memorizeCardSnapshot = await memorizeCard.get();
 
       await SubmissionProvider().use((provider) async {
-        await provider.setAllLocally(
-            submissionSnapshot.docs.map((e) => e.data()).toList());
-        eventBus.fire(SubmissionFetched());
+        await provider.writeTransaction(() async {
+          await provider.isar.clear();
+
+          await provider.putAllLocalOnly(submissionSnapshot.docs
+              .map((e) => Submission.fromMap(e.data()))
+              .toList());
+        });
       });
 
       await DigestiveProvider().use((provider) async {
-        await provider.setAllLocally(
-            digestiveSnapshot.docs.map((e) => e.data()).toList());
+        await provider.writeTransaction(() async {
+          await provider.putAllLocalOnly(digestiveSnapshot.docs
+              .map((e) => Digestive.fromMap(e.data()))
+              .toList());
+        });
       });
 
-      var timetableTables = timetableDataSnapshot.docs
-          .where((e) => e.id != "main")
-          .map((e) => {
-        "id": int.parse(e.id),
-        "title": e.data()["title"],
-      })
+      var timetableTables = timetableSnapshot.docs
+          .where((e) => e.id != "-1")
+          .map((e) => TimetableTable.fromMap({
+                "id": int.parse(e.id),
+                "title": e.data()["title"],
+              }))
           .toList();
 
+      print(timetableTables);
+
       await TimetableTableProvider().use((provider) async {
-        await provider.setAllLocally(timetableTables);
+        await provider.writeTransaction(() async {
+          await provider.putAllLocalOnly(timetableTables);
+        });
       });
 
       await TimetableProvider().use((provider) async {
-        await provider.deleteAllLocal();
-        for (var e in timetableDataSnapshot.docs) {
-          for (var value
-              in ((e.data()["cells"] as Map<String, dynamic>?) ?? {}).values) {
-            await provider.insertLocalOnly(provider.mapToObj(value));
+        await provider.writeTransaction(() async {
+          for (var e in timetableSnapshot.docs) {
+            var list = ((e.data()["cells"] as Map<String, dynamic>?) ?? {})
+                .values
+                .map((e) => Timetable.fromMap(e))
+                .toList();
+            await provider.putAllLocalOnly(list);
           }
-        }
+        });
       });
 
-      await TimetableClassTimeProvider(context).use((provider) async {
-        await provider.setAllLocally(
-            timetableClassTimeDataSnapshot.docs.map((e) => e.data()).toList());
+      await TimetableClassTimeProvider().use((provider) async {
+        await provider.writeTransaction(() async {
+          await provider.putAllLocalOnly(timetableClassTimeDataSnapshot.docs
+              .map((e) => TimetableClassTime.fromMap(e.data()))
+              .toList());
+        });
       });
 
-      await MemorizeCardFolderProvider().use((provider) async {
-        await provider.setAllLocally(memorizeCardSnapshot.docs
-            .map((e) => {
-                  "id": e.data()["id"],
-                  "title": e.data()["title"],
-                })
-            .toList());
-      });
+      // await MemorizeCardFolderProvider().use((provider) async {
+      //   await provider.setAllLocally(memorizeCardSnapshot.docs
+      //       .map((e) => {
+      //     "id": e.data()["id"],
+      //     "title": e.data()["title"],
+      //   })
+      //       .toList());
+      // });
 
       if (configData?.lastChanged != null) {
         SharedPrefs.use((prefs) {
@@ -264,93 +378,90 @@ class FirestoreProvider {
   }
 }
 
-class ConfigData {
-  ConfigData({
-    this.lastChanged,
+class UserConfig {
+  int? schemaVersion;
+  Timestamp? lastChanged;
+  Timestamp? lastAppOpened;
+  TimeOfDay? reminderNotificationTime;
+  TimeOfDay? timetableNotificationTime;
+  int? timetableNotificationId;
+  int? digestiveNotificationTimeBefore;
+  Lms? lms;
+
+  UserConfig({
     this.schemaVersion,
+    this.lastChanged,
+    this.lastAppOpened,
     this.reminderNotificationTime,
-    this.timetableNotification,
+    this.timetableNotificationTime,
+    this.timetableNotificationId,
     this.digestiveNotificationTimeBefore,
     this.lms,
   });
 
-  Timestamp? lastChanged;
-  int? schemaVersion;
-  TimeOfDay? reminderNotificationTime;
-  TimetableNotification? timetableNotification;
-  int? digestiveNotificationTimeBefore;
-  Lms? lms;
-
-  static ConfigData fromMap(Map<String, dynamic>? map) {
-    if (map == null) return ConfigData();
-    var reminderNotificationTimeSpilt =
-        (map["reminderNotificationTime"] as String?)?.split(":");
-    var timetableNotificationTimeSpilt =
-        (map["timetableNotificationTime"] as String?)?.split(":");
-    return ConfigData(
-      lastChanged: map["lastChanged"],
-      schemaVersion: map["schemaVersion"],
-      reminderNotificationTime: reminderNotificationTimeSpilt != null
-          ? TimeOfDay(
-              hour: int.parse(reminderNotificationTimeSpilt[0]),
-              minute: int.parse(reminderNotificationTimeSpilt[1]),
-            )
-          : null,
-      timetableNotification: TimetableNotification(
-        time: timetableNotificationTimeSpilt != null
-            ? TimeOfDay(
-                hour: int.parse(timetableNotificationTimeSpilt[0]),
-                minute: int.parse(timetableNotificationTimeSpilt[1]),
-              )
-            : null,
-        id: map["timetableNotificationId"],
-      ),
-      digestiveNotificationTimeBefore: map["digestiveNotificationTimeBefore"],
-      lms: Lms.fromMap(map["lms"]),
+  factory UserConfig.fromFirestore(DocumentSnapshot<Map<String, dynamic>> snapshot, SnapshotOptions? options) {
+    final data = snapshot.data();
+    return UserConfig(
+      schemaVersion: data?["schemaVersion"],
+      lastChanged: data?["lastChanged"],
+      lastAppOpened: data?["lastAppOpened"],
+      reminderNotificationTime: () {
+        final spilt = (data?["reminderNotificationTime"] as String?)?.split(":");
+        return spilt != null ? TimeOfDay(hour: int.parse(spilt[0]), minute: int.parse(spilt[1])) : null;
+      }(),
+      timetableNotificationTime: () {
+        final spilt = (data?["timetableNotificationTime"] as String?)?.split(":");
+        return spilt != null ?TimeOfDay(hour: int.parse(spilt[0]), minute: int.parse(spilt[1])) : null;
+      }(),
+      timetableNotificationId: data?["timetableNotificationId"],
+      digestiveNotificationTimeBefore: data?["digestiveNotificationTimeBefore"],
+      lms: Lms.fromMap(data?["lms"]),
     );
+  }
+
+  Map<String, dynamic> toFirestore() {
+    return {};
   }
 }
 
 class Lms {
-  CanvasLms? canvas;
+  Canvas? canvas;
 
-  Lms(Map<String, dynamic> map) : canvas = CanvasLms.fromMap(map["canvas"]);
+  Lms({this.canvas});
 
-  static fromMap(Map<String, dynamic>? map) {
-    if (map != null) {
-      return Lms(map);
-    }
-    return null;
+  static Lms? fromMap(Map<String, dynamic>? map) {
+    if (map == null) return null;
+    return Lms(
+      canvas: Canvas.fromMap(map["canvas"]),
+    );
   }
 }
 
-class CanvasLms {
-  int universityId;
+class Canvas {
+  int? universityId;
   Timestamp? lastSync;
-  bool hasError;
-  List<dynamic> excludedPlannableIds;
-  int submissionColor;
+  bool? hasError;
+  List<dynamic>? excludedPlannableIds;
+  int? submissionColor;
 
-  CanvasLms(Map<String, dynamic> map)
-      : universityId = map["universityId"],
-        lastSync = map["lastSync"],
-        hasError = map["hasError"],
-        excludedPlannableIds = map["excludedPlannableIds"],
-        submissionColor = map["submissionColor"];
+  Canvas({
+    this.universityId,
+    this.lastSync,
+    this.hasError,
+    this.excludedPlannableIds,
+    this.submissionColor,
+  });
 
-  static fromMap(Map<String, dynamic>? map) {
-    if (map != null) {
-      return CanvasLms(map);
-    }
-    return null;
+  static Canvas? fromMap(Map<String, dynamic>? map) {
+    if (map == null) return null;
+    return Canvas(
+      universityId: map["universityId"],
+      lastSync: map["lastSync"],
+      hasError: map["hasError"],
+      excludedPlannableIds: map["excludedPlannableIds"],
+      submissionColor: map["submissionColor"],
+    );
   }
-}
-
-class TimetableNotification {
-  TimetableNotification({this.time, this.id});
-
-  TimeOfDay? time;
-  int? id;
 }
 
 class SchemaVersionMismatchException {
