@@ -4,13 +4,14 @@ import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'package:intl/intl.dart';
 import 'package:submon/components/digestive_detail_card.dart';
 import 'package:submon/components/digestive_edit_bottom_sheet.dart';
-import 'package:submon/components/formatted_date_remaining.dart';
-import 'package:submon/db/digestive.dart';
-import 'package:submon/db/submission.dart';
+import 'package:submon/components/submissions/formatted_date_remaining.dart';
+import 'package:submon/isar_db/isar_digestive.dart';
+import 'package:submon/isar_db/isar_submission.dart';
 import 'package:submon/main.dart';
 import 'package:submon/utils/ui.dart';
-import 'package:submon/utils/utils.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:url_launcher/url_launcher_string.dart';
+
+import '../utils/ad_unit_ids.dart';
 
 class SubmissionDetailPage extends StatefulWidget {
   const SubmissionDetailPage(this.submissionId, {Key? key}) : super(key: key);
@@ -18,6 +19,7 @@ class SubmissionDetailPage extends StatefulWidget {
   final int submissionId;
 
   @override
+  // ignore: library_private_types_in_public_api
   _SubmissionDetailPageState createState() => _SubmissionDetailPageState();
 }
 
@@ -41,11 +43,8 @@ class _SubmissionDetailPageState extends State<SubmissionDetailPage> {
       });
     });
     DigestiveProvider().use((provider) async {
-      var digestives = await provider.getAll(
-          where: "$colSubmissionId = ?", whereArgs: [widget.submissionId]);
-      digestives.sort((a, b) {
-        return a.startAt.difference(b.startAt).inMinutes;
-      });
+      var digestives =
+          await provider.getDigestivesBySubmissionId(widget.submissionId);
       setState(() {
         _digestiveList = digestives;
       });
@@ -53,7 +52,7 @@ class _SubmissionDetailPageState extends State<SubmissionDetailPage> {
 
     if (!screenShotMode) {
       _bannerAd = BannerAd(
-        adUnitId: getAdUnitId(AdUnit.submissionDetailBanner)!,
+        adUnitId: AdUnits.submissionDetailBanner!,
         size: AdSize.banner,
         request: const AdRequest(),
         listener: const BannerAdListener(),
@@ -74,8 +73,8 @@ class _SubmissionDetailPageState extends State<SubmissionDetailPage> {
             onPressed: () async {
               await Navigator.pushNamed(context, "/submission/edit",
                   arguments: {"submissionId": widget.submissionId});
-              SubmissionProvider().use((provider) {
-                provider.get(widget.submissionId).then((value) {
+              SubmissionProvider().use((provider) async {
+                await provider.get(widget.submissionId).then((value) {
                   setState(() {
                     item = value;
                   });
@@ -86,8 +85,8 @@ class _SubmissionDetailPageState extends State<SubmissionDetailPage> {
         ],
       ),
       floatingActionButton: FloatingActionButton(
-        child: const Icon(Icons.add),
         onPressed: showCreateDigestiveBottomSheet,
+        child: const Icon(Icons.add),
       ),
       body: SafeArea(
         child: Column(
@@ -117,7 +116,7 @@ class _SubmissionDetailPageState extends State<SubmissionDetailPage> {
                             if (item != null)
                               Text(
                                   DateFormat("M/d (E)", 'ja_JP')
-                                      .format(item!.date!),
+                                      .format(item!.due),
                                   style: const TextStyle(fontSize: 18)),
                           ],
                         ),
@@ -131,8 +130,19 @@ class _SubmissionDetailPageState extends State<SubmissionDetailPage> {
                             const SizedBox(width: 8),
                             if (item != null)
                               FormattedDateRemaining(
-                                  item!.date!.difference(DateTime.now()),
+                                  item!.due.difference(DateTime.now()),
                                   numberSize: 24),
+                          ],
+                        ),
+                      ),
+                      if (item != null && item!.repeat != Repeat.none) Padding(
+                        padding: const EdgeInsets.all(8.0),
+                        child: Row(
+                          children: [
+                            const Spacer(),
+                            const Icon(Icons.repeat),
+                            const SizedBox(width: 8),
+                            Text(item!.repeat.toLocaleString()),
                           ],
                         ),
                       ),
@@ -143,13 +153,14 @@ class _SubmissionDetailPageState extends State<SubmissionDetailPage> {
                           // child: Text(item!.detail, style: const TextStyle(fontSize: 16, height: 1.3)),
                           child: Linkify(
                             onOpen: (link) async {
-                              if (await canLaunch(link.url)) {
-                                await launch(link.url);
+                              if (await canLaunchUrlString(link.url)) {
+                                await launchUrlString(link.url,
+                                    mode: LaunchMode.externalApplication);
                               } else {
                                 throw 'Could not launch $link';
                               }
                             },
-                            text: item!.detail,
+                            text: item!.details,
                             style: const TextStyle(fontSize: 16, height: 1.7),
                           ),
                         ),
@@ -189,9 +200,9 @@ class _SubmissionDetailPageState extends State<SubmissionDetailPage> {
             if (_bannerAd != null)
               Container(
                 alignment: Alignment.center,
-                child: AdWidget(ad: _bannerAd!),
                 width: _bannerAd!.size.width.toDouble(),
                 height: _bannerAd!.size.height.toDouble(),
+                child: AdWidget(ad: _bannerAd!),
               ),
           ],
         ),
@@ -207,7 +218,9 @@ class _SubmissionDetailPageState extends State<SubmissionDetailPage> {
     );
     if (data != null) {
       DigestiveProvider().use((provider) async {
-        await provider.insert(data);
+        provider.writeTransaction(() async {
+          await provider.put(data);
+        });
       });
       setState(() {
         _digestiveList.add(data);
